Consolidate graphql queries
